+++ lib/open_route_service.dart
@@ -9,6 +9,8 @@ export 'package:open_route_service/src/models/direction_data_models.dart';
 export 'package:open_route_service/src/models/elevation_data_models.dart';
 export 'package:open_route_service/src/models/geojson_feature_models.dart';
 export 'package:open_route_service/src/models/matrix_data_models.dart';
+export 'package:open_route_service/src/models/optimization_models/optimization_data_models.dart';
+export 'package:open_route_service/src/models/optimization_models/vroom_data_models.dart';
 export 'package:open_route_service/src/models/pois_data_models.dart';
 
 export 'package:open_route_service/src/open_route_service_base.dart';