+++ test/open_route_service_test.dart
@@ -4,6 +4,7 @@ import 'services/directions_tests.dart';
 import 'services/elevation_tests.dart';
 import 'services/isochrones_tests.dart';
 import 'services/matrix_tests.dart';
+import 'services/optimization_tests.dart';
 import 'services/pois_tests.dart';
 
 import 'package:open_route_service/open_route_service.dart';
@@ -21,6 +22,12 @@ Future<void> main() async {
       Coordinate(latitude: 49.41461, longitude: 8.681495);
   const Coordinate isochroneEndCoordinate =
       Coordinate(latitude: 49.41943, longitude: 8.686507);
+
+  // Dummy Optimization Jobs and Vehicles
+  const String serializedJobs =
+      '[{"id":1,"service":300,"amount":[1],"location":[1.98935,48.701],"skills":[1],"time_windows":[[32400,36000]]},{"id":2,"service":300,"amount":[1],"location":[2.03655,48.61128],"skills":[1]},{"id":3,"service":300,"amount":[1],"location":[2.39719,49.07611],"skills":[2]},{"id":4,"service":300,"amount":[1],"location":[2.41808,49.22619],"skills":[2]},{"id":5,"service":300,"amount":[1],"location":[2.28325,48.5958],"skills":[14]},{"id":6,"service":300,"amount":[1],"location":[2.89357,48.90736],"skills":[14]}]';
+  const String serializedVehicles =
+      ' [{"id":1,"profile":"driving-car","start":[2.35044,48.71764],"end":[2.35044,48.71764],"capacity":[4],"skills":[1,14],"time_window":[28800,43200]},{"id":2,"profile":"driving-car","start":[2.35044,48.71764],"end":[2.35044,48.71764],"capacity":[4],"skills":[2,14],"time_window":[28800,43200]}]';
 
   final OpenRouteService service = OpenRouteService(apiKey: apiKey);
   group('Initial test', () {
@@ -68,6 +75,15 @@ Future<void> main() async {
       service: service,
       boundingBoxStart: isochroneStartCoordinate,
       boundingBoxEnd: isochroneEndCoordinate,
+    ),
+  );
+
+  group(
+    'Optimization API tests:',
+    () => optimizationTests(
+      service: service,
+      serializedJobs: serializedJobs,
+      serializedVehicles: serializedVehicles,
     ),
   );
 }